pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "triagem-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlalchemy",
  "psycopg2-binary",
  "pydantic[email]",
  "typing-extensions",
  "passlib[bcrypt]",
  "PyJWT"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "httpx"
]

[tool.setuptools.packages.find]
include = ["backend*"]